resolved errors in pollcontoller... Configuration done
