+++ backend/src/main/java/com/example/backend/BackendApplication.java
@@ -1,12 +1,16 @@
 package com.example.backend;
 
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+
 
 @SpringBootApplication
 public class BackendApplication {
 
+
 	public static void main(String[] args) {
+
 		SpringApplication.run(BackendApplication.class, args);
 	}
