+++ backend/src/main/java/com/example/backend/config/URL.java
@@ -1,0 +1,6 @@
+package com.example.backend.config;
+
+public class URL {
+	public static final String BASE_URL = "/";
+	public static final String POLL = "/poll";
+}
