+++ backend/src/main/java/com/example/backend/entity/User.java
@@ -1,4 +1,4 @@
-package entity;
+package com.example.backend.entity;
 
 import java.util.List;
