+++ backend/src/main/java/com/example/backend/repository/UserRepository.java
@@ -1,9 +1,9 @@
-package repository;
+package com.example.backend.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import entity.User;
+import com.example.backend.entity.User;
 
 @Repository("userRepository")
 public interface UserRepository extends JpaRepository<User, Long> {
