+++ backend/src/main/java/com/example/backend/services/PollService.java
@@ -1,0 +1,8 @@
+package com.example.backend.services;
+
+import com.example.backend.entity.PollEntity;
+import com.example.backend.models.PollRequest;
+
+public interface PollService{
+	PollEntity createPoll(PollRequest pollRequest);
+}
