+++ backend/src/main/java/com/example/backend/services/UserService.java
@@ -1,0 +1,7 @@
+package com.example.backend.services;
+
+import com.example.backend.entity.User;
+
+public interface UserService {
+	public User getUserById(long id);
+}
