--- backend/src/main/java/config/URL.java
@@ -1,6 +1,0 @@
-package config;
-
-public class URL {
-	public static final String BASE_URL = "/";
-	public static final String POLL = "/poll/";
-}
