--- backend/src/main/java/services/PollService.java
@@ -1,8 +1,0 @@
-package services;
-
-import entity.PollEntity;
-import models.PollRequest;
-
-public interface PollService{
-	PollEntity createPoll(PollRequest pollRequest);
-}
