--- backend/src/main/java/services/UserService.java
@@ -1,7 +1,0 @@
-package services;
-
-import entity.User;
-
-public interface UserService {
-	public User getUserById(long id);
-}
