+++ backend/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/trouver
-spring.datasource.username=root
-spring.datasource.password=Gorzilla@1234
+spring.datasource.username=anony
+spring.datasource.password=Atul@123
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jackson.serialization.fail-on-empty-beans=false
